NO-ISS: Add missing depencies, and one missing #include (#5)
#### Why I did it

Build failed on AWS VM.

#### How I did it

Added missing dependencies into cmake config, and added a missing
#include where fmt::join was used. (I don't understand why this failed
on the VM but not on a local dev machine.)

#### How to verify it

Confirm that a build works. I've got some other stuff to commit before
you can easily do that.

#### Which release branch to backport

Is this relevant for the FBOSS work?

- [ ] 202411