NO-ISS: build_debug is not supported in FBOSS, and use main workflow (#15)
#### Why I did it

The logic for build_debug was broken, and ended up passing `-D` to the
fboss build_code_script.sh, which doesn't support that flag.

For experimenting I was using build.reusable.yml from a personal branch.
Now that has merged so we can use the main one.

#### How I did it

Just remove the argument altogether, because we don't do anything useful
with it in any case.

Change reference to build.reusable.yml.

#### How to verify it

This build passed with this change:
https://github.com/nexthop-ai/private-fboss/actions/runs/15055992315/job/42321914796

#### Which release branch to backport

- [ ] 202411