ISS-2121: unwedge PR title check GHA (#32)
#### Why I did it

After becoming a required check, the PR title check GHA occasionally
get's stuck. Adding the `synchronize` type to the `pull_request` event
seems to resolve the issue. (This is the same fix that was applied in
`nh` repo, here: https://github.com/nexthop-ai/nh/pull/1188)

#### How I did it
Add the `synchronize` activity type to the `pull_request` event.

#### How to verify it
These changes resolved the issue in `nh`, so this is expected to work.

#### Which release branch to backport

- [ ] 202411