ISS-2644 add support for reading the ONIE EEPROM format (#39)
#### Why I did it

FBOSS has its own EEPROM format, which we'll probably have to use when
shipping gear to Meta. However in our labs our devices will quickly
switch back and forth between SONiC and FBOSS, and it's not practical to
have to rewrite the EEPROM each time. It's better to have FBOSS support
the SONiC EEPROM format.

#### How I did it

Add support for loading the SONiC EEPROM format if that's the format we
detect.

#### How to verify it

Added unit tests.