ISS-2293 Add FBOSS agent config (#38)
#### Why I did it

We finally got the FBOSS SAI. This file is necessary to run, at least,
the SAI hw tests.

#### How I did it

I copied the Montblanc materialized config and changed the SDK
configuration for the NH-4010 128x400G config with the management ports
removed.

#### How to verify it

First build the FBOSS SAI, then link it with FBOSS using something like
the following commands:
```
./fboss/oss/scripts/build-helper.py sai/bcm_sai_external_fb/output/x86-xgsall-deb-static-fboss/libraries/libsai_impl.a sai/bcm_sai_external_fb/include/experimental /var/FBOSS/sai_impl 1.15.3
./fboss/oss/scripts/nhfboss-get-deps.sh
./fboss/oss/scripts/nhfboss-build.sh
```

With the included agent configuration file for NH-4010, we are able to
execute the SAI tests on the dut using the test runner:
```
sudo su
cd /opt/fboss
./bin/setup_fboss_env
./bin/run_test.py sai --config ~admin/nh4010.agent.materialized_JSON --sai-bin bin/sai_test-sai_impl 
```

#### Which release branch to backport

- [ ] 202411