NO-ISS: Add option to reset docker mounts (#57)
Also ran black

#### Why I did it

Since building the SAI and FBOSS against the SAI use the same container,
sometimes I need to reset its mount paths, such as for adding an
existing SAI or rerunning in a new directory. I would have to stop and
remove the docker container before rerunning the script with new mount
options.

#### How I did it

Make it more convenient to reset the mounts inside the container with a
new option `--reset-mounts`

I also ran black to quickly get rid of dangling whitespaces and it ended
up reconfiguring the rest of the script, which looks a lot more readable
-- anything below line 30 in the diff only has whitespace changes.

#### How to verify it

Ran the command with the --reset-mounts option both when the container
is already running and verified it changed the mounts and when the
container is not running to make sure it just no-ops.