NO-ISS: Add documentation to the nhfboss-* scripts (#65)
#### Why I did it

Give some clarity to the scripts, including possible notes on why the
script was created and any gotchas. This makes it faster to get a basic
understanding of the scripts without having to code read.

#### How I did it

Added comments. Comments for the nhfboss-build.py script are added in
https://github.com/nexthop-ai/private-fboss/pull/64

#### How to verify it

No behaviour changes, comments only.