ISS-3193 Support building FBOSS isolated in ephemeral build container (#58)
#### Why I did it

The FBOSS Image design adds an additional level of build system which
requires that each component, of which FBOSS itself is but one, must
build in an ephemeral build container.

Here that container image is created based on the pre-existing
Dockerfile and the necessary scripts to build the docker image, run a
command inside it, and a script to build FBOSS itself are added.

#### How I did it

#### How to verify it

#### Which release branch to backport

- [ ] 202411

---------

Co-authored-by: Anna Komarova <anna@nexthop.ai>