NO-ISS Commit latest stable hashes (#72)
#### Why I did it

The dependency pinning files which are committed do not include all the
dependencies. More dependencies are in
fboss/oss/stable_hashes/latest_stable_hashes.tar.gz.

Extract and commit those to maintain a working build instead of using
whatever is at the HEAD of the dependencies. Also stop .gitignoring the
build directory so we can see the changes there.

This may also stabilize the cached build times since we would no longer
be at the whims of the upstream dependencies making commits and
invalidating our cache.

#### How I did it

#### How to verify it

#### Which release branch to backport

- [ ] 202411