ISS-3194 Bazel build FBOSS with Fake SAI (#67)
#### Why I did it

Here a bazel target which builds FBOSS against the fake SAI is added.

This is a stepping-stone to running the FBOSS unit tests as part of the
build

As with building the [FBOSS
SAI](https://github.com/nexthop-ai/broadcom-sai-sdk/pull/15) we
explicitly leave the intermediate build products in the source tree to
improve incremental rebuild speed.

#### How I did it

#### How to verify it

#### Which release branch to backport

- [ ] 202411