NO-ISS Share the server during FBOSS build  (#77)
#### Why I did it

Reduce the parallelism of the FBOSS build to leave 2 CPUs for other
uses. Also nice the build itself to allow others to use the CPUs; the
FBOSS build will be so long that a few minutes of sharing doesn't
matter.

Hopefully this makes building FBOSS on dev servers less disruptive to
other users.

#### How I did it

#### How to verify it

#### Which release branch to backport

- [ ] 202411