NO-ISS: Create per-user non-root FBOSS build containers (#89)
#### Why I did it

We don't actually need to do everything as `root` inside the container,
and doing things as `root` causes other issues such as creating files in
the `.git` directory as `root` that can then no longer be modified
outside of the container.

#### How I did it

Pass the outer username, UID, GID into the Docker build in order to
create a user inside the container that is almost identical to the user
outside the container. This is similar to what SONiC does. The change
remain backwards compatible as the Dockerfile defaults to `root` if no
user is specified.