ISS-3556: FBOSS Platform Name sanitize function for NH4010 (#79)
#### Why I did it

Platform Name for Golden Eagle is listed as `NH-4010` in bios and as
`NH4010` in code some place. We need to write a sanitize function to
translate between them to make some tests not fail because of this
difference.

#### How I did it

- Sanitize Platform Name (e.g. `NH-4010` to `NH4010`)

#### How to verify it

- Without this change `weutil_hw_test` and `data_corral_service_hw_test`
failed at the start

- With this change `weutil_hw_test` and `data_corral_service_hw_test`
are able to move past platform name check


#### Which release branch to backport

- [ ] 202411