NO-ISS Move FBOSS build dir under fboss checkout (#88)
#### Why I did it

This change moves the build scratch directory under the fboss source
tree. It has turned out not to be necessary for the scratch directory to
be outside the FBOSS directory and placing it directly under /work does
not maintain a desirable amount of separate between clones because often
the first build is performed with no branch and therefore shares the
same directory across all clones.

Instead place the build directory at fboss/.build_dir so it follows the
clone.

#### How I did it

#### How to verify it

#### Which release branch to backport

- [ ] 202411