NO-ISS: Ensure gcc-11 isn't installed in the FBOSS container (#106)
#### Why I did it

My recent changes in #89 broke things for out-of-the-box bash users. I
tried to make it such that both bash and zsh users could source the
right file to use gcc-12. But really the correct solution is to fix the
`Dockerfile` to not install any other version of GCC in the first place.

#### How I did it

Don’t install the package group `Development Tools`, install all the
packages it contains except `gcc` because we instead install
`gcc-toolset-12`.

#### How to verify it

Did a full build+test cycle in a clean container.