NO-ISS: New GHA to build and test FBOSS nightly (#108)
#### Why I did it

So we can build and test our private fork for each PR and also nightly and catch issues.

#### How I did it

Create a new VM on cloudstack called fboss-gha-runner, configured a GHA
runner on it, added a new workflow script to run the build using a
prebuild SAI. We preserve the pinned dependencies across builds to speed
up incremental builds.

#### How to verify it

Manual GHA trigger.