NO-ISS: Use remote sccache workers (#124)
#### Why I did it

Local FBOSS builds are slow except on a huge machine.

#### How I did it

Use the the remote sccache workers that Brad set up.

sccache has a behavior where it runs a build locally if the remote
workers are busy. This is would bring a machine to its knees if we’re
building with parallelism set for the remote workers and the full load
ends up on the local machine. I made a patched sccache that will simply
keep retrying the remote build in that case. We install *that* sccache
in the Dockerfile now. See that PR at
https://github.com/timn-nexthop/sccache/pull/1

If the sccache scheduler is reachable, configure sccache to use it and
use high parallelism for the build. Otherwise, pick some numbers based
on how much RAM the server his.

On a VM with 12 cores and 24GB of memory, this ran nhfboss-build.sh
(with no cache hits) in 20m4s.

#### How to verify it

#### Which release branch to backport

- [ ] 202411

<!-- GitButler Footer Boundary Top -->
---
This is **part 1 of 2 in a stack** made with GitButler:
- <kbd>&nbsp;2&nbsp;</kbd> #127 
- <kbd>&nbsp;1&nbsp;</kbd> #124 👈 
<!-- GitButler Footer Boundary Bottom -->