NO-ISS: Set the OOM score adjustment more broadly (#141)
# Summary

Set the OOM score adjustment more broadly

Set the `oom_score_adj` in the common script so that it not only applies
to our build script but also the get-deps stage and unit tests.

Also reduce the number of `sccache` jobs that we run at once as it was 
sometimes overwhelming the system, with the `sccache` server itself 
taking 10G of RAM.

# Test Plan

Existing tests pass.