NO-ISS: Allow passing a timeout to unit tests (#140)
# Summary

Allow passing a timeout to unit tests

When running unit tests, no timeout was enforced. Add an optional 
`--timeout` flag to `getdeps.py` that is passed to `ctest` and update 
our build script to set a default timeout of 90 seconds per test, which 
is generous enough for now.

# Test Plan

Existing unit tests pass within the given timeout.