NO-ISS: Enable Meta's pre-commit hook (#158)
# Summary

I disabled it when I first merged it becuase I was wrestling with issues
in files that were getting flagged by this hook that weren’t caused by 
us, but we should really enable this so that we don’t introduce lint 
errors that get flagged when we upstream PRs.

# Test Plan

Pre-commit hook should pass from now on.