NOS-510: Remove calling undefined function (#165)
**Pre-submission checklist**
- [x] I've ran the linters locally and fixed lint errors related to the
files I modified in this PR. You can install the linters by running `pip
install -r requirements-dev.txt && pre-commit install`
- [x] `pre-commit run`

# Summary

Forgot to remove calling a function that I removed and did not catch it
in my testing since the test fails on fboss101 with autodetecting board
being enabled, so it exited before hitting this line. Discovered when
running passing tests on gold222

# Test Plan

retested on gold222 to verify it can now print out the summary instead
of throwing an error.