NO-ISS: Build missing test binaries automatically when running tests (#170)
# Summary

When building only a specific cmake target to save time, getdeps removes
the "built marker” under `$builddir/installed/fboss/.built-by-getdeps`, 
which makes it impossible to subsequently run the unit tests as  without
this file getdeps bails out with `project fboss has not been built`. 
What’s more, because of a missing `return` the exit code was 0, which 
was confusing.
This fix ensures that the return code is propagated properly and also 
improves the cmake test runner to automatically detect missing 
executables and build them on the fly. This will not rebuild the test 
binaries if they've already been built and are merely stale (because the
source code has changed) but that was already the prior behavior. At 
least now it's possible to rebuild a single cmake test target and re-run
the tests in one shot.

# Test Plan

Not needed.