NO-ISS: Don't try to run the pre-commit hook on sync commits (#182)
# Summary

When we sync from upstream, we pull hundreds of commits touching all
sorts of files, some of which don’t pass the pre-commit hook (!). So we
get yelled at when we sync, but it’s not our fault, it’s not even our
changes!
This changes the pre-commit to ignore sync merge commits.

# Test Plan

Works on the latest sync PR.