NOS-360 Build Distro Image in CI (#162)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please provide enough information so
that others can review your pull request. -->

**Pre-submission checklist**
- [ ] I've ran the linters locally and fixed lint errors related to the
files I modified in this PR. You can install the linters by running `pip
install -r requirements-dev.txt && pre-commit install`
- [ ] `pre-commit run`

# Summary

<!-- Explain the motivation for making this change and any other context
that you think would help reviewers of your code. What existing problem
does the pull request solve? -->

This is the start of the Distro Image CI where the Distro Image is
built. Eventually this will be merged upstream to build these images in
the FBOSS OSS Github repo.

As part of this it is necessary to remove `-it` from the docker command
because GHA's don't have a stdin. Further the sample image manifest now
requires all components be present, so add them.

# Test Plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran
and their output, screenshots / videos if the pull request changes the
user interface. How exactly did you verify that your PR solves the issue
you wanted to solve? -->

<!-- If a relevant Github issue exists for this PR, please make sure you
link that issue to this PR -->

Look at the new workflow in this PR.