NOS-3267: Build and test PRs with clang only. (#233)
# Summary

We've been double-building and testing all PRs with GCC and clang for
exactly one month (since 11/24). There were no major issues except for
a couple little escapes caught early on. The whole team has switched
to clang, and Meta is actively reviewing the clang changes to try to
upstream them. So let's just switch our CI/CD to clang-only. The risk
is that we introduce a change that breaks the GCC build, then we won't
find out until Meta tells us when they import our PR that there is a
problem with it. It's not very likely but it's possible.


# Test Plan

We’ve all already been testing this for a month.